+++ argopy/data_fetchers/localftp_data.py
@@ -121,20 +121,6 @@ class LocalFTPArgoDataFetcher(ArgoDataFetcherProto):
         summary.append("Domain: %s" % self.cname())
         return '\n'.join(summary)
 
-    def _format(self, x, typ):
-        """ string formating helper """
-        if typ == 'lon':
-            if x < 0:
-                x = 360. + x
-            return ("%05d") % (x * 100.)
-        if typ == 'lat':
-            return ("%05d") % (x * 100.)
-        if typ == 'prs':
-            return ("%05d") % (np.abs(x) * 10.)
-        if typ == 'tim':
-            return pd.to_datetime(x).strftime('%Y%m%d')
-        return str(x)
-
     def _absfilepath(self, wmo: int, cyc: int = None, errors: str = 'raise') -> str:
         """ Return the absolute netcdf file path to load for a given wmo/cyc pair
 
@@ -445,7 +431,7 @@ class Fetch_wmo(LocalFTPArgoDataFetcher):
 class Fetch_box(LocalFTPArgoDataFetcher):
     """ Manage access to local ftp Argo data for: a rectangular space/time domain  """
 
-    def init(self, box: list = [-180, 180, -90, 90, 0, 6000, '1900-01-01', '2100-12-31']):
+    def init(self, box: list ):
         """ Create Argo data loader
 
             Parameters
@@ -458,10 +444,19 @@ class Fetch_box(LocalFTPArgoDataFetcher):
         # but we work only with x, y and t.
         if len(box) not in [6, 8]:
             raise ValueError('Box must be 6 or 8 length')
+
         if len(box) == 6:
             self.BOX = [box[ii] for ii in [0, 1, 2, 3]]
         elif len(box) == 8:
             self.BOX = [box[ii] for ii in [0, 1, 2, 3, 6, 7]]
+
+        # if len(box) == 6:
+        #     # Select the last months of data:
+        #     end = pd.to_datetime('now')
+        #     start = end - pd.DateOffset(months=1)
+        #     self.BOX.append(start.strftime('%Y-%m-%d'))
+        #     self.BOX.append(end.strftime('%Y-%m-%d'))
+
         self.fs_index = indexstore(self.cache, self.cachedir, os.path.sep.join([self.local_ftp, "ar_index_global_prof.txt"]))
         # todo we would need to check if the index file is there
